Player Resources - Implemented a new system for player Resouces using a Manager class to manager all the player resources (did it this way for easier access once networking comes into play)
Research
- Fixed some research animations that wasnt fading out when a card was selected. the state for those cards wasn't being changed.

TO DO:
Visual
- Need a new Radius UI using shaders on the world environment.

Networking
- Learn and Add networking function through client to server arch.

Units
- Problem with the animation events not being registered in code (Setup Animation function)
- Need to pass the data of the units when they get created in the unitpoolmanager.
- Fix the unit class, units data values dont reset when they die in game.
- Change the UI for the units.

- Finish off the rts changes to the unit movement and attacking.

+++ Assets/_Scripts/Scriptable/UnitScriptable.cs
@@ -35,8 +35,8 @@
 
         [Header("DATA - ATTRIBUTE VALUES")]
         public int turnCost = 1; // The Numbner of turns it takes for it to be able to spawn.
-        public float goldCost = 500.0f;
-        public float populationCost = 1.0f;
+        public int goldCost = 500;
+        public int populationCost = 1;
         [Range(0.0f, 3.0f)] public float resistanceMultiplier = 0.0f;
         [Range(0.0f, 3.0f)] public float weaknessMultiplier = 0.0f;
         public float heath = 0.0f;
